[ADD] Fix csv sale order uos qty to compute margin
When import sale order data from CSV the onchanges that works in the instance
view ae not called.

In the Instance view, when add a new line and you update the quantity field
(product_uom_qty) the product_uos_qty is updated with the same value, this
value even if product_uos_qty is not visible for the user.

The CSV set the product_uom_qty in the line,  but do not set the
product_uos_qty, When product_uos_qty is not set then the sale order line
margin is not properly compute.

In order to properly compute the sale order margin and then the sale order
margin we need to set the product_uos_qty of this sales.